requirements.txt
flask
groq
python-dotenv
supabase
gunicorn
flask-cors
